backend/requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
pydantic>=2.9.0
asyncpg>=0.29.0
qdrant-client>=1.7.0
openai>=1.10.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
